requirements.txt
flask==3.0.0
faiss-cpu==1.7.4
numpy==1.24.3
requests==2.31.0
sentence-transformers==2.2.2
tqdm==4.66.1

